9C-432 Unify PriceFilter Enumeration
We unify the definitions of the PriceFilter enumeration,
and apply them in the messages and methods across the app
