--- modules/api/src/main/scala/cards/nine/api/messages/PathEnumerations.scala
@@ -1,17 +1,0 @@
-package cards.nine.api.messages
-
-import enumeratum.{ Enum, EnumEntry }
-
-object PathEnumerations {
-
-  sealed trait PriceFilter extends EnumEntry
-
-  object PriceFilter extends Enum[PriceFilter] {
-    case object ALL extends PriceFilter
-    case object FREE extends PriceFilter
-    case object PAID extends PriceFilter
-
-    val values = super.findValues
-  }
-
-}
