+++ modules/api/src/main/scala/cards/nine/api/utils/SprayMatchers.scala
@@ -1,8 +1,7 @@
 package cards.nine.api.utils
 
-import cards.nine.api.messages.PathEnumerations.PriceFilter
 import cards.nine.domain.analytics.{ Continent, Country }
-import cards.nine.domain.application.Category
+import cards.nine.domain.application.{ Category, PriceFilter }
 import enumeratum.{ Enum, EnumEntry }
 import shapeless._
 import spray.http.Uri.Path
