+++ modules/commons/src/main/scala/cards/nine/domain/Application.scala
@@ -1,5 +1,7 @@
 package cards.nine.domain.application
 
+import enumeratum.{ Enum, EnumEntry }
+
 /**
   * A Package is the unique identifier of an android App.
   * It is a dot-separated sequence of lowercase segments, much like Java packages.
@@ -30,3 +32,13 @@ case class FullCardList(
   missing: List[Package],
   cards: List[FullCard]
 )
+
+sealed trait PriceFilter extends EnumEntry
+object PriceFilter extends Enum[PriceFilter] {
+  case object ALL extends PriceFilter
+  case object FREE extends PriceFilter
+  case object PAID extends PriceFilter
+
+  val values = super.findValues
+}
+
